More of an exploratory commit.  Feel free to revert if this is not a productive addition!
Established Routing rule and pattern for having App navigate to a Mentor view page.  (e.g. /Mentor/234 will load the Mentor view and can grab the shirt number as a possible key)

Changed the Main.view.xml's ColumnListItem type to Navigation and wrote a press handler to perform the routing.